fix: 中国退伍军人事务局无法获取到正文 (#3142)
