+++ lib/router.js
@@ -1422,10 +1422,8 @@ router.get('/maoyan/upcoming', require('./routes/maoyan/upcoming'));
 // cnBeta
 router.get('/cnbeta', require('./routes/cnbeta/home'));
 
-// 退伍士兵信息
-router.get('/gov/veterans/bnxx', require('./routes/gov/veterans/bnxx'));
-router.get('/gov/veterans/zcjd', require('./routes/gov/veterans/zcjd'));
-router.get('/gov/veterans/index', require('./routes/gov/veterans/index'));
+// 国家退伍士兵信息
+router.get('/gov/veterans/:type', require('./routes/gov/veterans/china'));
 
 // 河北省退伍士兵信息
 router.get('/gov/veterans/hebei/:type', require('./routes/gov/veterans/hebei'));
